Continued separations of models.py
